pyproject.toml
[project]
authors = [
    {name = "Your Name", email = "you@example.com"},
]
requires-python = "<4.0,>=3.12"
dependencies = [
    "yfinance<1.0.0,>=0.2.54",
    "pydantic<3.0.0,>=2.10.6",
    "tweepy[async]<5.0.0,>=4.15.0",
    "kafka-python<3.0.0,>=2.0.5",
    "loguru<1.0.0,>=0.7.3",
    "pyspark<4.0.0,>=3.5.5",
    "ruff>=0.9.9,<1.0.0",
    "transformers<5.0.0,>=4.50.0",
    "nltk<4.0.0,>=3.9.1",
    "emoji<3.0.0,>=2.14.1",
    "cassandra-driver<4.0.0,>=3.29.2",
    "polygon-api-client<2.0.0,>=1.14.4",
    "praw<8.0.0,>=7.8.1",
    "python-dotenv<2.0.0,>=1.1.0",
    "vadersentiment>=3.3.2",
    "typer>=0.15.2",
    "pendulum>=3.0.0",
    "confluent-kafka>=2.9.0",
    "pymongo>=4.12.0",
    "streamlit>=1.44.1",
    "plotly>=6.0.1",
    "pandas>=2.2.3",
    "fastapi>=0.115.12",
    "uvicorn>=0.34.2",
    "motor>=3.7.0",
]
name = "marketpulseai"
version = "0.1.0"
description = ""
readme = "README.md"
